% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cafe-Interacción con Miro.R
\name{cond_indepOnvar}
\alias{cond_indepOnvar}
\title{This function extracts the subset of the implied conditional
independences that containe an specified node name.}
\usage{
cond_indepOnvar(indeps, var)
}
\arguments{
\item{indeps}{Frame data as recovered from the Miro board.}

\item{var}{target variable to filter the conditional independence list.}
}
\value{
string containing the subset of statements of conditional independence.
}
\description{
This function extracts the subset of the implied conditional
independences that containe an specified node name.
}
