% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cafe-Interacción con Miro.R
\name{datosMiro}
\alias{datosMiro}
\title{Get Miro data into R}
\usage{
datosMiro(servMiro = "miro", user)
}
\arguments{
\item{servMiro}{Name of the credential service as defined in keyring setup}

\item{user}{User name as defined n keyring setup}
}
\value{
a list of nodes, arcs and frames attributes
}
\description{
This function accesses data from a specified Miro board describing
the fundamental elements of a Bayesian network: nodes and arcs.
Sticky notes are assumed to represent nodes, and connectors are the arcs.
Nodes are expected to have one and only one label attached. This label
is used to indicate the variable name that corresponds to the node in
the database.
It uses the "keyring library" to deal with credentials. It is expected
that credentials are prepared in advance with keyring::key_set(
service = servMiro, username = user). You should provide the token
produced by Miro when prompted for it in a popup window.
}
