% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miro2bayes.R
\name{miro2bnlearn}
\alias{miro2bnlearn}
\title{This function feeds the data from Miro into a bnlearn network.}
\usage{
miro2bnlearn(variables, arcs)
}
\arguments{
\item{variables}{Node data as recovered from Miro board}

\item{var}{Node data as recovered from Miro board.}
}
\value{
tibble::tibble with numbers sumirizing the network structure.
}
\description{
This function feeds the data from Miro into a bnlearn network.
}
