% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miro2bayes.R
\name{miro2DNE}
\alias{miro2DNE}
\title{This funnction builds a representation of the Bayesian
Network recovered from Miro in a DNE file format that
Netica and other Bayesian Network application can reed.}
\usage{
miro2DNE(miroData)
}
\arguments{
\item{frames_data}{Frame data as recovered from Miro board}

\item{variables}{Node data as recovered from Miro board}

\item{arcos}{Arc data as recovered from Miro board}
}
\value{
a string that contains the whole DNE document
}
\description{
This funnction builds a representation of the Bayesian
Network recovered from Miro in a DNE file format that
Netica and other Bayesian Network application can reed.
}
