% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miro2bayes.R
\name{miro_validar}
\alias{miro_validar}
\title{This function does a quick check on the nmeric consistency
of the network attributes as interpreted from the data
found in the Miro board.}
\usage{
miro_validar(papelitos, arcos)
}
\arguments{
\item{papelitos}{Node data as recovered from Miro board}

\item{var}{Node data as recovered from Miro board.}
}
\value{
tibble::tibble with numbers sumirizing the network structure.
}
\description{
This function does a quick check on the nmeric consistency
of the network attributes as interpreted from the data
found in the Miro board.
}
