% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miro2bayes.R
\name{miroBoards}
\alias{miroBoards}
\title{List all Miro board}
\usage{
miroBoards(servMiro = "miro", user)
}
\arguments{
\item{servMiro}{Name of the credential service as defined in keyring setup}

\item{user}{User name as defined n keyring setup}
}
\description{
This function recovers a list of all the boards available to the user
identified by provided credentials.
}
