% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miro2bayes.R
\name{miroValidation}
\alias{miroValidation}
\title{This function does a quick check on the numeric consistency
of the network attributes as interpreted from the data
found in the Miro board.}
\usage{
miroValidation(miroData)
}
\arguments{
\item{miroData}{Data recovered from Miro board}
}
\value{
tibble::tibble with numbers sumirizing the network structure.
}
\description{
This function does a quick check on the numeric consistency
of the network attributes as interpreted from the data
found in the Miro board.
}
