% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cafe-Interacción con Miro.R
\name{prepara_DAG}
\alias{prepara_DAG}
\title{Assemble the DAG as recovered from Miro sticky notes and arcs}
\usage{
prepara_DAG(nodos, arcos)
}
\arguments{
\item{nodos}{Node data as recovered from Miro board}

\item{arcos}{Arc data as recovered from Miro board}
}
\value{
a list holding a ggdag plot, text list of
implied conditional independenceof nodes, arcs and frames attributes,
and the DAG itself as interpreted by dagitty.
}
\description{
This function builds a formal DAG sing DOT language and dagitty
Using dagitty it produces a list of implied conditional independence conditions.
Using ggdag pprodces a graphical representation of the DAG.
}
