% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cafe-Interacción con Miro.R
\name{red_DNE}
\alias{red_DNE}
\title{This funnction builds a representation of the Bayesian
Network recovered from Miro in a DNE file format that
Netica and other Bayesian Network application can reed.}
\usage{
red_DNE(datos_marcos, papelitos, arcos)
}
\arguments{
\item{datos_marcos}{Frame data as recovered from Miro board}

\item{papelitos}{Node data as recovered from Miro board}

\item{arcos}{Arc data as recovered from Miro board}
}
\value{
a string that contains the whole DNE document
}
\description{
This funnction builds a representation of the Bayesian
Network recovered from Miro in a DNE file format that
Netica and other Bayesian Network application can reed.
}
