% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cafe-Interacción con Miro.R
\name{datosMiro}
\alias{datosMiro}
\title{Get Miro data into R}
\usage{
datosMiro(servMiro = "miro", user = "miguel-token")
}
\arguments{
\item{servMiro}{Name of the credential service as defined in keyring setup}

\item{user}{User name as defined n keyring setup}
}
\value{
a list of nodes, arcs and frames attributes
}
\description{
This function access data from an specified Miro board
describing the basic element of a Bayesiian Network: nodes and arcs
sticky notes are assumed to represent nodes and connectors the arcs among them.
Nodes are expected to have only one label attached to indicate the variable name that
that corresponds to the node in the database. It uses keyring for credentiials.
}
