% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miro2bayes.R
\name{miro2bnlearn}
\alias{miro2bnlearn}
\title{This function feeds the data from Miro into a bnlearn network.}
\usage{
miro2bnlearn(nodes, arcs, frames)
}
\arguments{
\item{arcs}{Data on arcs links as recovered from Miro board.}

\item{variables}{Data on Nodes as recovered from Miro board}
}
\value{
tibble::tibble with numbers sumirizing the network structure.
}
\description{
This function feeds the data from Miro into a bnlearn network.
}
